requirements.txt
aiohttp==3.1.3
amqp==5.0.6
asgiref==3.4.1
async-timeout==2.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
HackerNews==2.0.0
haxor==1.2.3
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==4.7.6
openapi-codec==1.3.2
packaging==21.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
uritemplate==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
yarl==1.7.0
